Classes de serviço para os parâmetros

+++ justebeer-app/src/main/java/br/com/lucasbibianot/security/SecurityJWT.java
@@ -79,7 +79,7 @@ public class SecurityJWT {
 		jwt.parseJWT(token);
 	}
 
-	public String createJWT(String login, int ttl) {
+	public String createJWT(String login, Long ttl) {
 		return this.createJWT(login, "localHost", login, ttl);
 	}
 }
